Configured signin and signup functions to one react component and updated CSS for signin page
