+++ frontend/food-app/src/components/auth/SignIn.css
@@ -1,9 +1,66 @@
+/* SignIn.css */
 .Login-text {
-    text-align: center
+    text-align: center;
 }
 
 .Login-prompt {
     display: flex;
-    align-items: center;  
+    align-items: center;
     justify-content: center;
+    height: 100vh; /* Adjust the height as needed */
 }
+
+/* Form styles */
+.Login-text label {
+    font-size: 18px;
+    font-weight: bold;
+    display: block;
+    margin-top: 10px;
+}
+
+.Login-text input {
+    width: 100%;
+    padding: 10px;
+    margin: 5px 0;
+    border: 1px solid #ccc;
+    border-radius: 5px;
+    font-size: 16px;
+    outline: none;
+}
+
+.Login-text button {
+    width: 100%;
+    padding: 10px;
+    margin: 10px 0;
+    background-color: #007BFF; /* Adjust the background color as needed */
+    color: #fff;
+    border: none;
+    border-radius: 5px;
+    font-size: 18px;
+    cursor: pointer;
+    transition: background-color 0.3s;
+}
+
+.Login-text button:hover {
+    background-color: #0056b3; /* Adjust the hover color as needed */
+}
+
+.Login-text .psw a {
+    text-decoration: none;
+    color: #007BFF; /* Adjust the link color as needed */
+}
+
+.Login-text .psw a:hover    {
+    text-decoration: underline;
+}
+
+/* Error message style */
+.error {
+    color: #FF0000; /* Red color for error messages */
+    font-size: 16px;
+    margin-top: 10px;
+}
+
+
+  
+
